+++ source/Messaging.Application/Transactions/MoveIn/MasterDataDelivery/ReceiveCustomerMasterData.cs
@@ -17,7 +17,7 @@ using System.Text.Json.Serialization;
 using Messaging.Application.Configuration.Commands.Commands;
 using Messaging.Application.MasterData;
 
-namespace Messaging.Application.Transactions.MoveIn;
+namespace Messaging.Application.Transactions.MoveIn.MasterDataDelivery;
 
 public class ReceiveCustomerMasterData : InternalCommand
 {
