+++ source/Messaging.IntegrationTests/Application/Transactions/MoveIn/MasterDataDelivery/ReceiveCustomerMasterDataTests.cs
@@ -18,7 +18,7 @@ using MediatR;
 using Messaging.Application.Configuration.DataAccess;
 using Messaging.Application.MasterData;
 using Messaging.Application.OutgoingMessages.CharacteristicsOfACustomerAtAnAp;
-using Messaging.Application.Transactions.MoveIn;
+using Messaging.Application.Transactions.MoveIn.MasterDataDelivery;
 using Messaging.Domain.Transactions.MoveIn;
 using Messaging.Infrastructure.Configuration.DataAccess;
 using Messaging.Infrastructure.Configuration.Serialization;
@@ -26,7 +26,7 @@ using Messaging.IntegrationTests.Fixtures;
 using NodaTime;
 using Xunit;
 
-namespace Messaging.IntegrationTests.Application.Transactions.MoveIn;
+namespace Messaging.IntegrationTests.Application.Transactions.MoveIn.MasterDataDelivery;
 
 public class ReceiveCustomerMasterDataTests
     : TestBase, IAsyncLifetime
